+++ src/pallet/extensions.clj
@@ -1,8 +1,15 @@
 (ns pallet.extensions
-  (:use [clojure.contrib.def :only (name-with-attributes)])
+  (:use [clojure.tools.macro :only (name-with-attributes symbol-macrolet)])
   (:require pallet.resource.filesystem-layout
-            [clojure.contrib.condition :as condition]
-            [clojure.contrib.macro-utils :as macro]))
+            [clojure.tools.macro :as macro]))
+
+;; pallet 0.6 uses slingshot 0.5.x which is in the slinghshot.core
+;; namespace, whereas 0.7+ uses slingshot 0.10.x which is in the
+;; slingshot.slingshot namespace. Good thing this is not Java!
+(try
+  (use '[slingshot.slingshot :only [throw+]])
+  (catch Exception _
+    (use '[slingshot.core :only [throw+]])))
 
 ;; ### Pallet Extensions
 ;;
@@ -29,7 +36,7 @@
   expressions. Future iterations will include more symbol macro
   bindings."
   [& forms]
-  `(macro/symbol-macrolet
+  `(symbol-macrolet
     [~'when pallet.thread-expr/when->
      ~'for pallet.thread-expr/for->
      ~'let pallet.thread-expr/let->
@@ -63,20 +70,20 @@
   [session form]
   (if (and session (map? session))
     session
-    (condition/raise
-     :type :invalid-session
-     :message
-     (str
-      "Invalid session map in phase.\n"
-      (format "session is %s\n" session)
-      (format "Problem probably caused by subphase:\n  %s\n" form)
-      "Check for non crate functions, improper crate functions, or
-      problems in threading the session map in your phase
-      definition. A crate function is a function that takes a session
-      map and other arguments, and returns a modified session
+    (throw+ {
+             :type :invalid-session
+             :message
+             (str
+              "Invalid session map in phase.\n"
+              (format "session is %s\n" session)
+              (format "Problem probably caused by subphase:\n  %s\n" form)
+              "Check for non crate functions, improper crate
+      functions, or problems in threading the session map in your
+      phase definition. A crate function is a function that takes a
+      session map and other arguments, and returns a modified session
       map. Calls to crate functions are often wrapped in a threading
       macro, -> or pallet.phase/phase-fn, to simplify chaining of the
-      session map argument."))))
+      session map argument.")})))
 
 (defmacro phase-fn
   "Composes a phase function from a sequence of phases by threading an