+++ ReleaseNotes.md
@@ -1,4 +1,8 @@
 # Release Notes
+# 0.7.0-beta.1
+
+- Add support for pallet 0.6 + clojure 1.2.x and pallet 0.7 + clojure 1.3+
+- Default to pallet 0.7.0
 
 # 0.7.0-alpha.2